src/back/requirements.txt
fastapi
uvicorn
mysqlclient
SQLAlchemy
sqlalchemy_utils
alembic
PyMySQL
cryptography
python-dotenv
requests_oauthlib
python-dateutil